test: replace SleepingRunner with MockEventProcessorRunner
